Moved location of components
Placed footer, header and navbar into separate folder called navigation and updated references to components to match new relative path.
