+++ src/components/Navigation/Header.jsx
@@ -1,4 +1,4 @@
-import Navigation from "./Navigation";
+import Navigation from "./NavigationBar";
 
 function Header() {
     return (
