+++ src/components/Navigation/NavigationBar.jsx
@@ -2,9 +2,9 @@ import Container from 'react-bootstrap/Container';
 import Nav from 'react-bootstrap/Nav';
 import Navbar from 'react-bootstrap/Navbar';
 import NavDropdown from 'react-bootstrap/NavDropdown';
-import logo from '../logoPrototype.png';
+import logo from '../../logoPrototype.png';
 
-function Navigation() {
+function NavigationBar() {
   return (
     <Navbar bg="light" expand="lg">
       <Container>
@@ -47,4 +47,4 @@ function Navigation() {
   );
 }
 
-export default Navigation;
+export default NavigationBar;
